+++ llmfoundry/command_utils/__init__.py
@@ -1,0 +1,11 @@
+# Copyright 2024 MosaicML LLM Foundry authors
+# SPDX-License-Identifier: Apache-2.0
+from llmfoundry.command_utils.eval import (
+    eval_from_yaml,
+    evaluate,
+)
+
+__all__ = [
+    'evaluate',
+    'eval_from_yaml',
+]
