+++ tests/a_scripts/eval/test_eval_inputs.py
@@ -8,7 +8,7 @@ import pytest
 from omegaconf import DictConfig
 from omegaconf import OmegaConf as om
 
-from llmfoundry.train import evaluate  # noqa: E402
+from llmfoundry.command_utils import evaluate  # noqa: E402
 
 
 class TestHuggingFaceEvalYAMLInputs:
